+++ src/adapters/MorphoBlueAdapter.sol
@@ -16,7 +16,7 @@ import {MathLib} from "../libraries/MathLib.sol";
 struct PositionInMarket {
     uint128 shares;
     uint128 assets;
-    uint pendingDeposits;
+    uint256 pendingDeposits;
 }
 
 contract MorphoBlueAdapter is IMorphoBlueAdapter {
@@ -75,7 +75,7 @@ contract MorphoBlueAdapter is IMorphoBlueAdapter {
 
     /// @dev Does not log anything because the ids (logged in the parent vault) are enough.
     /// @dev Returns the ids of the allocation and the change in assets in the position.
-    function allocate(bytes memory data, uint assets) external returns (bytes32[] memory, int256) {
+    function allocate(bytes memory data, uint256 assets) external returns (bytes32[] memory, int256) {
         MarketParams memory marketParams = abi.decode(data, (MarketParams));
         require(msg.sender == parentVault, NotAuthorized());
         require(marketParams.loanToken == asset, LoanAssetMismatch());
@@ -91,7 +91,7 @@ contract MorphoBlueAdapter is IMorphoBlueAdapter {
         uint256 mintedShares = position.pendingDeposits.toSharesDown(totalSupplyAssets, totalSupplyShares);
         if (mintedShares > 0) {
             position.shares += uint128(mintedShares);
-            (uint depositedAssets, ) = IMorpho(morpho).supply(marketParams, 0, mintedShares, address(this), hex"");
+            (uint256 depositedAssets,) = IMorpho(morpho).supply(marketParams, 0, mintedShares, address(this), hex"");
             position.pendingDeposits -= depositedAssets;
         }
         uint256 newAssetsInMarket = expectedSupplyAssets(marketParams, position.shares);
